Parser/requirements.txt
flask
opencv-python
pytesseract
pdf2image
pillow
numpy
pyzbar
lxml
